Engine

See merge request CompSci308_2018Spring/voogasalad_callussalad!262

+++ src/game_engine/components/HealthComponent.java
@@ -2,8 +2,8 @@ package game_engine.components;
 
 import game_engine.Component;
 
-public class HealthComponent extends Component<Integer> {
+public class HealthComponent extends Component<Double> {
 	public HealthComponent(String arg){
-		super(Integer.parseInt(arg));
+		super(Double.parseDouble(arg));
 	}
 }
